pyproject.toml
[project]
name = "AYU-CSE-BACKEND"
version = "0.0.1"
description = "Anyang University CSE"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]==3.2.7",
    "pydantic-settings==2.9.1",
    "sqlalchemy>=2.0.40",
]
